backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
redis==5.0.1
rq==1.13.0
gunicorn==21.2.0
pytest==7.4.0
requests==2.31.0
flask-cors==4.0.1
bcrypt==4.0.1
alembic==1.12.0
